# Minimize the calls in useSelect by subscribing to only the stores needed