--- a/packages/data/src/components/use-select/index.js
+++ b/packages/data/src/components/use-select/index.js
@@ -37,6 +37,23 @@ const useIsomorphicLayoutEffect =
 const renderQueue = createQueue();
 
 /**
+ * Traverse up the registries to find the registry that has the store registered.
+ *
+ * @param {Object} registry The registry obtained by `useRegistry`.
+ * @param {string} storeKey The store key name.
+ * @return {Object | null} The registered store or null if it doesn't exists in the registry tree.
+ */
+function getRegisteredStore( registry, storeKey ) {
+	if ( ! registry ) {
+		return null;
+	} else if ( storeKey in registry.stores ) {
+		return registry.stores[ storeKey ];
+	}
+
+	return getRegisteredStore( registry.parent, storeKey );
+}
+
+/**
  * Custom react hook for retrieving props from registered selectors.
  *
  * In general, this custom React hook follows the
@@ -192,11 +209,10 @@ export default function useSelect( _mapSelect, deps ) {
 		};
 
 		const unsubscribers = listeningStores.current
-			// The registry could be a sub-registry that doesn't have the store directly.
-			// Skip subscribing to the store that doesn't exist.
-			.filter( ( storeKey ) => storeKey in registry.stores )
-			.map( ( storeKey ) =>
-				registry.stores[ storeKey ].subscribe( onChange )
+			.map( ( storeKey ) => getRegisteredStore( registry, storeKey ) )
+			.filter( ( registeredStore ) => registeredStore !== null )
+			.map( ( registeredStore ) =>
+				registeredStore.subscribe( onChange )
 			);
 		const unsubscribe = () => {
 			unsubscribers.map( ( unsubscriber ) => unsubscriber() );