--- a/packages/data/src/registry.js
+++ b/packages/data/src/registry.js
@@ -223,6 +223,7 @@ export function createRegistry( storeConfigs = {}, parent = null ) {
 	}
 
 	let registry = {
+		parent,
 		registerGenericStore,
 		stores,
 		namespaces: stores, // TODO: Deprecate/remove this.