--- a/packages/data/src/registry.js
+++ b/packages/data/src/registry.js
@@ -12,6 +12,30 @@ import createCoreDataStore from './store';
 
 /** @typedef {import('./types').WPDataStore} WPDataStore */
 
+// A map to record each registry and their stores/parent privately.
+const StoresMap = new WeakMap();
+
+/**
+ * Traverse up the registries to find the registry that has the store registered.
+ *
+ * @param {Object} registry  The registry created by createRegistry below.
+ * @param {string} storeName The store key name.
+ * @return {Object | null} The registered store or null if it doesn't exists in the registry tree.
+ */
+function getRegisteredStore( registry, storeName ) {
+	if ( ! registry ) {
+		return null;
+	}
+
+	const { stores, parent } = StoresMap.get( registry );
+
+	if ( storeName in stores ) {
+		return stores[ storeName ];
+	}
+
+	return getRegisteredStore( parent, storeName );
+}
+
 /**
  * @typedef {Object} WPDataRegistry An isolated orchestrator of store registrations.
  *
@@ -222,8 +246,25 @@ export function createRegistry( storeConfigs = {}, parent = null ) {
 		registerGenericStore( store.name, store.instantiate( registry ) );
 	}
 
+	/**
+	 * Subscribe handler to all the stores.
+	 *
+	 * @param {string[]} listeningStores Store names in array.
+	 * @param {Function} handler         The function subscribed to the stores.
+	 * @return {Function} A function to unsubscribe every handler.
+	 */
+	function __experimentalSubscribeStores( listeningStores, handler ) {
+		const unsubscribers = listeningStores
+			.map( ( storeName ) => getRegisteredStore( registry, storeName ) )
+			.filter( ( registeredStore ) => registeredStore !== null )
+			.map( ( registeredStore ) => registeredStore.subscribe( handler ) );
+
+		return () => {
+			unsubscribers.forEach( ( unsubscriber ) => unsubscriber() );
+		};
+	}
+
 	let registry = {
-		parent,
 		registerGenericStore,
 		stores,
 		namespaces: stores, // TODO: Deprecate/remove this.
@@ -234,6 +275,7 @@ export function createRegistry( storeConfigs = {}, parent = null ) {
 		use,
 		register,
 		__experimentalMarkListeningStores,
+		__experimentalSubscribeStores,
 	};
 
 	/**
@@ -279,5 +321,16 @@ export function createRegistry( storeConfigs = {}, parent = null ) {
 		parent.subscribe( globalListener );
 	}
 
-	return withPlugins( registry );
+	const storesMap = {
+		stores,
+		parent,
+	};
+	StoresMap.set( registry, storesMap );
+
+	const registryWithPlugins = withPlugins( registry );
+
+	// We don't need to do this once `withPlugins` is removed.
+	StoresMap.set( registryWithPlugins, storesMap );
+
+	return registryWithPlugins;
 }