@adamziel 2020-11-05
@kevin940726 it doesn't account for selectors like

```js
export const getWidgets = createRegistrySelector( ( select ) => () => {
	const widgets = select( 'core' ).getEntityRecords(
		'root',
		'widget',
		buildWidgetsQuery()
	);

	return keyBy( widgets, 'id' );
} );
```

To make it work, you'd need to somehow hook into actual `select( store )`, store all the synchronous calls, and access them in `useSelect`. I wonder it that could get somehow mixed up with an unrelated `mapSelect` too.