@adamziel 2020-11-05
@kevin940726 I played with it more and came up with this: https://github.com/WordPress/gutenberg/pull/26733 - still not working properly though