@youknowriad 2020-11-10
Does it support this as well https://github.com/WordPress/gutenberg/pull/26724#issuecomment-722330948 ?