@adamziel 2020-11-23
@kevin940726 I think this PR is obsolete now that https://github.com/WordPress/gutenberg/pull/26866 was merged