@youknowriad 2020-11-23
Actually, I'm considering reverting the `@wordpress/stan` temporarily (to give me time to do more performance tests as I discovered some issues that need clarification).  Maybe we can land this smaller approach after the revert.