@youknowriad 2020-11-26
It seems the performance impact here is good but not as good as what we had in the atom's approach. 
I believe the difference with atoms, is that there was no need for the selectors to run again inside `useEffect` on first mount. This optimization though, doesn't seem to be possible here.