@adamziel 2020-11-27
> I believe the difference with atoms, is that there was no need for the selectors to run again inside useEffect on first mount. This optimization though, doesn't seem to be possible here.

Our stores are large so it's normal to have selectors "watching" a branch that stays the same, that's another difference with the atoms approach - only relevant changes bubble up to their observers.